Fix configuration path when outside project directory
When running the plugin from an editor, the current working directory may not be the same as the project directory. Filepath never seems to be passed in as an option, so just fake it with the extension and resolve the directory name to the project directory this is running in.
Note: this assumes that this plugin is installed in a node_modules folder of a project, not globally.